python/requirements.txt
pipecat-ai==0.0.25
openai-whisper==20231117
piper-tts==1.2.0
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.0
librosa==0.10.1
soundfile==0.12.1
fastapi==0.104.0
uvicorn==0.24.0
python-multipart==0.0.6
pytest
